prepare reuse of DelegatingApp
